refactor(appsec): slim down appsec index file
