+++ packages/dd-trace/test/appsec/sdk/user_blocking-integration.spec.js
@@ -10,7 +10,7 @@ const appsec = require('../../../src/appsec')
 const blocking = require('../../../src/appsec/blocking')
 const { getConfigFresh } = require('../../helpers/config')
 const agent = require('../../plugins/agent')
-const { json } = require('../../../src/appsec/blocked_templates')
+const { json } = require('../../../src/appsec/blocking/templates')
 
 describe('user_blocking - Integration with the tracer', () => {
   const config = getConfigFresh({
